+++ macos/real_liquid_glass/Package.swift
@@ -2,19 +2,19 @@
 import PackageDescription
 
 let package = Package(
-    name: "liquid_glass_container",
+    name: "real_liquid_glass",
     platforms: [
         .macOS("10.14")
     ],
     products: [
-        .library(name: "liquid-glass-container", targets: ["liquid_glass_container"])
+        .library(name: "real-liquid-glass", targets: ["real_liquid_glass"])
     ],
     dependencies: [
         .package(name: "FlutterMacOSFramework", path: "../FlutterMacOSFramework")
     ],
     targets: [
         .target(
-            name: "liquid_glass_container",
+            name: "real_liquid_glass",
             dependencies: [
                 .product(name: "FlutterMacOSFramework", package: "FlutterMacOSFramework")
             ]